requirements.txt
mcp[cli]>=1.0.0
pyyaml>=6.0
pyathena>=3.0.0
boto3>=1.26.0
